requirements.txt
fastapi==0.115.8
sqlalchemy==2.0.41
pymysql==1.1.1
pytest==8.4.0
python-dotenv==1.1.0
passlib==1.7.4
bcrypt==4.3.0
